Fix crashing from cleaning the extract dir (#1)
If the extraction directory doesn't exist then Files#walk will crash trying to get the attributes of a non-existent path

+++ src/main/java/vazkii/cmpdl/CMPDL.java
@@ -215,16 +215,18 @@ public final class CMPDL {
     public static void extractModpackMetadata(Path zipFile, Path extractDir) throws IOException {
         Interface.setStatus("Unzipping Modpack Download");
 
-        log("Cleaning up extract dir");
-        Files.walk(extractDir)
-                .sorted(Comparator.reverseOrder())
-                .forEachOrdered(path -> {
-                    try {
-                        Files.deleteIfExists(path);
-                    } catch (IOException e) {
-                        throw new UncheckedIOException(e);
-                    }
-                });
+        if (Files.exists(extractDir)) {
+            log("Cleaning up extract dir");
+            Files.walk(extractDir)
+                    .sorted(Comparator.reverseOrder())
+                    .forEachOrdered(path -> {
+                        try {
+                            Files.deleteIfExists(path);
+                        } catch (IOException e) {
+                            throw new UncheckedIOException(e);
+                        }
+                    });
+        }
 
         log("Unzipping file");
         try (FileSystem zipFs = FileSystems.newFileSystem(zipFile, (ClassLoader) null)) {
